+++ client/gui/InterfaceObjectConfigurable.h
